113 - move style file to core/style
