+++ lib/features/runs/presentation/ui/widgets/place_run_button.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:hotfoot/src/utils/style.dart';
+import 'package:hotfoot/core/style/style.dart';
 
 class PlaceRunBotton extends StatelessWidget {
   final VoidCallback _onPressed;
