+++ lib/features/runs/presentation/ui/widgets/run_card.dart
@@ -5,7 +5,7 @@ import 'package:hotfoot/features/navigation_screen/presentation/bloc/navigation_
 import 'package:hotfoot/features/runs/data/models/run_model.dart';
 import 'package:hotfoot/features/runs/domain/entities/run_entity.dart';
 import 'package:hotfoot/features/runs/domain/entities/run_status.dart';
-import 'package:hotfoot/src/utils/style.dart';
+import 'package:hotfoot/core/style/style.dart';
 import 'package:intl/intl.dart';
 
 class RunCard extends StatelessWidget {
