+++ lib/features/runs/presentation/ui/widgets/run_form.dart
@@ -6,7 +6,7 @@ import 'package:hotfoot/features/places/domain/entities/place_entity.dart';
 import 'package:hotfoot/features/runs/presentation/ui/widgets/place_name.dart';
 import 'package:hotfoot/features/runs/presentation/ui/widgets/place_run_button.dart';
 import 'package:hotfoot/features/runs/presentation/ui/widgets/run_photo.dart';
-import 'package:hotfoot/src/utils/style.dart';
+import 'package:hotfoot/core/style/style.dart';
 
 class RunForm extends StatefulWidget {
   final PlaceEntity placeEntity;
